+++ tests/atcVersionsClient.unit.tests.ts
@@ -34,17 +34,21 @@ describe('atc versions unit tests', function () {
 
     it('refresh atc versions', async function () {
 
-        nock('https://api.github.com:443')
-            .get('/repos/F5Networks/f5-appsvcs-templates/releases')
-            .reply(200, fastResp)
-            .get('/repos/F5Networks/f5-appsvcs-extension/releases')
-            .reply(200, as3Resp)
-            .get('/repos/F5Networks/f5-declarative-onboarding/releases')
-            .reply(200, doResp)
-            .get('/repos/F5Networks/f5-telemetry-streaming/releases')
-            .reply(200, tsResp)
-            .get('/repos/F5Networks/f5-cloud-failover-extension/releases')
-            .reply(200, cfResp)
+
+        // *** enable the following code block to nock all responses ***
+        //      This is only needed when repeatedly running tests and possibly hitting github api quota limit
+
+        // nock('https://api.github.com:443')
+        //     .get('/repos/F5Networks/f5-appsvcs-templates/releases')
+        //     .reply(200, fastResp)
+        //     .get('/repos/F5Networks/f5-appsvcs-extension/releases')
+        //     .reply(200, as3Resp)
+        //     .get('/repos/F5Networks/f5-declarative-onboarding/releases')
+        //     .reply(200, doResp)
+        //     .get('/repos/F5Networks/f5-telemetry-streaming/releases')
+        //     .reply(200, tsResp)
+        //     .get('/repos/F5Networks/f5-cloud-failover-extension/releases')
+        //     .reply(200, cfResp)
 
         const extHttp = new ExtHttp();
         const atcV = new AtcVersionsClient({